handling pools (#252)
* handling pools

* cleanup

* redeployed RelayPoolNativeGateway

* adding types

* adding mainnet deployments

* deployed gateway on ethereum

* fixing script to add origins

* wip

* deployed Bridge Factory on Lisk

* deployed BridgeProxy for op on Ethereum

* deployed BridgeProxy on Lisk

* updated task to bridge

* wip

* adding call to issue insufficient gas

* wip

* adding withgdrawal thru timelock

* adding a script to withdraw from a pool

* cleanup

* redeployed factory

* deployed a new pool

* deployed OP BridgeProxy on ethereum

* deployed an op bridge on lisk

* adding missing address

* updated packages

* more script updates

* linting

* more changes

* more changes

* LoadEmitted handled

* updated versions

+++ smart-contracts/contracts/RelayPool.sol
@@ -475,9 +475,10 @@ contract RelayPool is ERC4626, Ownable {
     if (amount > 0) {
       updateStreamedAssets();
       // We adjust the end of the stream based on the new amount
-      uint amountLeft = remainsToStream();
-      uint timeLeft = Math.max(endOfStream, block.timestamp) - block.timestamp;
-      uint weightedStreamingPeriod = (amountLeft *
+      uint256 amountLeft = remainsToStream();
+      uint256 timeLeft = Math.max(endOfStream, block.timestamp) -
+        block.timestamp;
+      uint256 weightedStreamingPeriod = (amountLeft *
         timeLeft +
         amount *
         streamingPeriod) / (amountLeft + amount);
